We (@Ampholyt and me) have implemented the algorithm from https://stackoverflow.com/a/59432406/12540220. We had to make some changes to the way food_sources are stored in a Dot.
We also changed it so, that a Dot can only eat one food per tick at max and not every food below it. This might explain the (small) changes in the attributes of the Dots at the end, if they aren't just a result of the random mechanics. Overall the result ist pretty amazing:
Current Master:
```
--- Werte von 305 Dots in Tick 100000 ---
        | MIN | MAX | AVG |
SIGHT:  |  15 |  70 |  30 |
SPEED:  |  8  |  33 |  16 |
ENERGY: |  34 |14031| 3358|
Genaue Sightwerte:
Wert:   | 15|| 17| 18| 19| 20| 21| 22| 23| 24| 25| 26| 27| 28| 29| 30| 31| 32| 3
3| 34| 35| 36| 37| 38| 39| 40| 41| 42| 43| 44| 45| 46| 47|| 49| 50| 51| 52| 53|
54| 55||| 58||| 61| 62|||||||| 70|
Anzahl: | 1 || 6 | 2 | 11| 11| 15| 16| 17| 17| 13| 17| 4 | 19| 10| 17| 14| 12| 7
 | 6 | 5 | 8 | 6 | 8 | 8 | 7 | 2 | 3 | 6 | 2 | 4 | 2 | 5 || 5 | 1 | 2 | 4 | 3 |
1 | 2 ||| 1 ||| 3 | 1 |||||||| 1 |
Genaue Speedwerte:
Wert:   | 8 | 9 | 10| 11| 12| 13| 14| 15| 16| 17| 18| 19| 20| 21| 22| 23| 24| 25
| 26| 27| 28| 29|||| 33|
Anzahl: | 1 | 6 | 6 | 20| 24| 27| 36| 24| 27| 17| 18| 9 | 19| 23| 11| 8 | 14| 6
| 1 | 4 | 1 | 2 |||| 1 |
Im aktuellen Tick sind 876 Futterstuecke auf dem Feld
Zeit 992.344
```
New search algorithm
```
--- Werte von 274 Dots in Tick 100000 ---
        | MIN | MAX | AVG |
SIGHT:  |  19 |  69 |  34 |
SPEED:  |  6  |  27 |  15 |
ENERGY: |  1  |17097| 3682|
Genaue Sightwerte:
Wert:   | 19| 20| 21| 22| 23| 24| 25| 26| 27| 28| 29| 30| 31| 32| 33| 34| 35| 36
| 37| 38| 39| 40| 41| 42| 43| 44| 45| 46| 47| 48| 49| 50| 51| 52| 53| 54| 55| 56
| 57| 58| 59||| 62| 63| 64|| 66||| 69|
Anzahl: | 4 | 10| 13| 4 | 12| 11| 17| 9 | 13| 12| 7 | 15| 5 | 12| 10| 14| 6 | 7
| 8 | 3 | 8 | 6 | 5 | 4 | 7 | 4 | 5 | 4 | 1 | 2 | 4 | 3 | 1 | 3 | 1 | 4 | 1 | 3
| 5 | 2 | 2 ||| 1 | 3 | 1 || 1 ||| 1 |
Genaue Speedwerte:
Wert:   | 6 | 7 | 8 | 9 | 10| 11| 12| 13| 14| 15| 16| 17| 18| 19| 20| 21| 22| 23
| 24| 25| 26| 27|
Anzahl: | 1 | 3 | 16| 10| 8 | 20| 14| 19| 24| 13| 18| 19| 22| 15| 20| 15| 13| 8
| 6 | 4 | 4 | 2 |
Im aktuellen Tick sind 935 Futterstuecke auf dem Feld
Zeit 287.201
```
This is a speedup of around x3.46!

Also changed getFunctions to const. This is usually standard.