@IceFreez3r 2020-07-07
It seems like, the above times were measured without compiling with `-O3`. If we add the flag, both times go down, but the times of master way more. This results in more similar times in the end:
Master
```
--- Werte von 299 Dots in Tick 100000 ---
        | MIN | MAX | AVG |
SIGHT:  |  16 |  71 |  31 |
SPEED:  |  7  |  32 |  17 |
ENERGY: |  14 |13203| 3200|
Genaue Sightwerte:
Wert:   | 16| 17| 18| 19| 20| 21| 22| 23| 24| 25| 26| 27| 28| 29| 30| 31| 32| 33
| 34| 35| 36| 37| 38| 39| 40| 41| 42| 43| 44|| 46| 47| 48| 49| 50|| 52||||||| 59
|| 61| 62| 63||| 66| 67| 68| 69|| 71|
Anzahl: | 4 | 2 | 4 | 6 | 7 | 7 | 13| 17| 16| 18| 21| 20| 7 | 10| 8 | 16| 17| 13
| 8 | 8 | 5 | 12| 5 | 5 | 8 | 4 | 4 | 4 | 4 || 4 | 3 | 1 | 5 | 2 || 1 ||||||| 1
|| 1 | 1 | 1 ||| 1 | 1 | 2 | 1 || 1 |
Genaue Speedwerte:
Wert:   | 7 | 8 | 9 | 10| 11| 12| 13| 14| 15| 16| 17| 18| 19| 20| 21| 22| 23| 24
| 25| 26| 27| 28| 29| 30| 31| 32|
Anzahl: | 2 | 4 | 8 | 8 | 11| 20| 22| 23| 17| 21| 13| 19| 20| 10| 12| 20| 14| 13
| 11| 13| 8 | 3 | 3 | 2 | 1 | 1 |
Im aktuellen Tick sind 853 Futterstuecke auf dem Feld
Zeit 81.03
```
SearchFood Grid
```
--- Werte von 288 Dots in Tick 100000 ---
        | MIN | MAX | AVG |
SIGHT:  |  18 |  93 |  36 |
SPEED:  |  6  |  32 |  15 |
ENERGY: |  18 |16319| 3457|
Genaue Sightwerte:
Wert:   | 18| 19| 20| 21| 22| 23| 24| 25| 26| 27| 28| 29| 30| 31| 32| 33| 34| 35
| 36| 37| 38| 39| 40| 41| 42| 43| 44| 45|| 47| 48| 49| 50| 51| 52| 53| 54| 55| 5
6| 57| 58| 59|| 61| 62| 63| 64||| 67| 68|||| 72|| 74|| 76|| 78||| 81|| 83|||||||
|| 92| 93|
Anzahl: | 3 | 1 | 3 | 12| 10| 13| 13| 13| 13| 9 | 9 | 11| 13| 7 | 9 | 9 | 7 | 9
| 13| 11| 11| 11| 9 | 4 | 3 | 7 | 3 | 4 || 2 | 1 | 1 | 4 | 1 | 2 | 1 | 3 | 1 | 2
 | 2 | 4 | 2 || 1 | 1 | 3 | 1 ||| 3 | 5 |||| 1 || 1 || 1 || 1 ||| 1 || 1 |||||||
|| 1 | 1 |
Genaue Speedwerte:
Wert:   | 6 | 7 | 8 | 9 | 10| 11| 12| 13| 14| 15| 16| 17| 18| 19| 20| 21| 22| 23
| 24| 25| 26| 27|||| 31| 32|
Anzahl: | 1 | 2 | 9 | 11| 13| 18| 20| 37| 24| 16| 20| 20| 21| 14| 15| 11| 11| 9
| 4 | 3 | 4 | 3 |||| 1 | 1 |
Im aktuellen Tick sind 862 Futterstuecke auf dem Feld
Zeit 78.122
```
But since the times are still reliably lower, we will still merge.